requirements.txt
atari-py==0.2.6
box2d-py==2.3.8
cffi==1.13.2
cloudpickle==1.2.2
Cython==0.29.14
fasteners==0.15
future==0.18.2
glfw==1.8.5
gym==0.15.4
imageio==2.6.1
lockfile==0.12.2
monotonic==1.5
numpy==1.17.4
opencv-python==4.1.2.30
pandas==0.25.3
Pillow==6.2.1
pycparser==2.19
pyglet==1.3.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
scipy==1.3.3
six==1.13.0
torchvision==0.4.2
